Changes to the way plugins are stored internally. Plugins are now indexed by their plugin_id rather than their filename, though the plugin_id is largely influenced by filename. The InfoObject class was added to extend InfoRecord as a read-only equivalent.  This is used to store plugin info (metadata) so that magic __get() and __set() need not be overridden in descendant plugin classes. The Plugins admin page is modified to accommodate the changes, including some possible silly code to make the plugin list a 1/3 size column so that the plugin config forms can appear above the fold.  CSS is yet to be applied to any of this to make it pretty, someone please do that kthxbye.
Thanks Freakerz, for the initial patch.

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@714 653ae4dd-d31e-0410-96ef-6bf7bf53c507

+++ classes/inforecords.php
@@ -75,6 +75,12 @@ abstract class InfoRecords implements URLProperties
 			$this->_loaded == TRUE;
 			return;
 		}
+		
+		// This InfoRecord is read-only?
+		if(empty($this->_table_name)) {
+			$this->_loaded == TRUE;
+			return;
+		}
 		 
 		$result= DB::get_results( '
 			SELECT name, value, type
@@ -174,6 +180,10 @@ abstract class InfoRecords implements URLProperties
 	 **/	
 	public function delete_all()
 	{
+		// This InfoRecord is read-only?
+		if(empty($this->_table_name)) {
+			return;
+		}
 		$result= DB::query( '
 			DELETE FROM ' . $this->_table_name . '
 			WHERE ' . $this->_key_name . ' = ?',
